Conflicts:
	src/main/java/br/org/gdt/beans/CsbffBeneficioBean.java
	src/main/java/br/org/gdt/beans/CsbffDadosPessoaisBean.java
	src/main/java/br/org/gdt/beans/CsbffDadosProfissionaisBean.java
	src/main/java/br/org/gdt/dao/RecPessoaDAO.java
	src/main/java/br/org/gdt/model/RecPessoa.java
	src/main/webapp/c_s_b_f_f/admissao/dadosprofissionais.xhtml
	src/main/webapp/c_s_b_f_f/admissao/listaadmissao.xhtml
	src/main/webapp/c_s_b_f_f/beneficio/consultabeneficios.xhtml
	src/main/webapp/resources/build/js/utils.js

+++ src/main/java/br/org/gdt/dao/GchPerguntasDAO.java
@@ -38,7 +38,7 @@ public class GchPerguntasDAO extends DAO<GchPerguntas> {
         }
     }
 
-    public List<GchPerguntas> buscaPergutasFormulario(int id) {
+    public List<GchPerguntas> buscaPergutasFormulario(long id) {
 
         Query query = entityManager.createQuery("from GchPerguntas as t where t.formulario.formCodigo = :codigo");
         query.setParameter("codigo", id);
